rename sql_encoding
